Change implementation to use request headers.
Russ Fenn mentioned that the MicroStrategy documentation seems to indicate
that real request headers are checked for the requisite authn/authz data.
Note this implementation is fundamentally less secure since request headers
are under the control of the client. Russ suggested an Apache pipeline
component that strips out the headers if present prior to authentication to
address that concern.